+++ lib/datarockets_style/cop/rspec/prefer_before_over_setup.rb
@@ -1,0 +1,39 @@
+# frozen_string_literal: true
+
+module DatarocketsStyle
+  module Cop
+    module RSpec
+      # bad
+      # setup do
+      #   allow(post).to receive(:publish!)
+      # end
+      #
+      # good
+      # before do
+      #   allow(post).to receive(:publish!)
+      # end
+      class PreferBeforeOverSetup < RuboCop::Cop::Cop
+        MSG = "Prefer `before` instead of `setup`."
+
+        def_node_matcher :setup_call?, <<-PATTERN
+          (block
+            (send _ :setup)
+            (args) _)
+        PATTERN
+
+        def on_block(node)
+          return unless setup_call?(node)
+
+          add_offense(node)
+        end
+
+        def autocorrect(node)
+          lambda do |corrector|
+            block_internals = node.source.split(/ /)
+            corrector.replace node.loc.expression, ["before", *block_internals[1..-1]].join(" ")
+          end
+        end
+      end
+    end
+  end
+end